+++ examples/pytorch_track.py
@@ -11,7 +11,7 @@ device = torch.device('cpu')
 # Hyper parameters
 num_epochs = 5
 num_classes = 10
-batch_size = 50
+batch_size = 32
 learning_rate = 0.01
 
 # aim - Track hyper parameters